refactor: extract useDarkMode hook from App.jsx
Move dark mode state (darkMode, useOsTheme, systemPrefersDark),
the appearanceDark useMemo, and the matchMedia/data-bs-theme
useEffects into a self-contained useDarkMode hook.
